% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_S3.R
\name{calculate_limit}
\alias{calculate_limit}
\title{"Detection" limit for each model}
\usage{
calculate_limit(model, pars, limit, temp_range)
}
\arguments{
\item{model}{character string defining the inactivation model to use.}

\item{pars}{list defining the model parameters according to the rules defined in the bioinactivation package.}

\item{limit}{numerical value describing the maximum number of log-reductions
that can be identified in the experiment limit = logDL - logN0, where DL
is the detection limit.}

\item{temp_range}{Numeric vector that defines the range of possible temperatures}
}
\value{
Numerical value that indicates the limit of detection
}
\description{
Calculation of the detection limit depending on the model.
}
