% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM_calculation.R
\name{criterium_D}
\alias{criterium_D}
\title{D Optimality Criterium}
\usage{
criterium_D(FIM)
}
\arguments{
\item{FIM}{Matrix with the values of the Fisher Information Matrix}
}
\description{
D Optimality Criterium
}
