% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_limits.R
\name{get_detection}
\alias{get_detection}
\title{Calculate detection limit}
\usage{
get_detection(model, pars, temperature, limit)
}
\arguments{
\item{model}{character string defining the inactivation model to use.}

\item{pars}{list defining the model parameters according to the rules defined in the bioinactivation package.}

\item{temperature}{numerical value that describes the temperature at which the detection limit will be calculated}

\item{limit}{numerical value describing the maximum number of log-reductions
that can be identified in the experiment limit = logDL - logN0, where DL
is the detection limit.}
}
\value{
Numerical value that indicates the limit of detection
}
\description{
Calculation of the detection limit depending on the model.
}
